Crate: move grid related property to grid mod
