+++ takumi/src/style/properties/mod.rs
@@ -9,6 +9,8 @@ pub mod box_shadow;
 pub mod color;
 /// Gap properties for flex and grid layouts.
 pub mod gap;
+/// Grid-related properties (columns/rows/track sizing)
+pub mod grid;
 /// Length units and measurement types for the takumi styling system.
 pub mod length_unit;
 /// Linear gradient properties for elements.
@@ -25,18 +27,17 @@ pub use length_unit::*;
 pub use linear_gradient::*;
 pub use parser::*;
 pub use sides::*;
+// grid module re-exports grid types
+pub use grid::*;
 
 use cosmic_text::{Align, FamilyOwned, Weight};
 use cssparser::{BasicParseError, ParseError, Parser};
 use image::imageops::FilterType;
 use serde::{Deserialize, Serialize};
-use taffy::{
-  CompactLength, GridTemplateRepetition, MaxTrackSizingFunction, MinTrackSizingFunction,
-  TrackSizingFunction,
-};
+// Grid-specific taffy types moved to `properties::grid`
 use ts_rs::TS;
 
-use crate::{core::viewport::RenderContext, impl_from_taffy_enum};
+use crate::impl_from_taffy_enum;
 
 type ParseResult<'i, T, E = BasicParseError<'i>> = Result<T, ParseError<'i, E>>;
 
@@ -319,195 +320,6 @@ impl From<FontFamily> for FamilyOwned {
       FontFamily::Cursive => FamilyOwned::Cursive,
       FontFamily::Fantasy => FamilyOwned::Fantasy,
       FontFamily::Custom(name) => FamilyOwned::Name(name.into()),
-    }
-  }
-}
-
-/// Represents a grid track sizing function with serde support
-#[derive(Debug, Clone, Deserialize, Serialize, TS, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum GridLengthUnit {
-  /// A fraction of the available space
-  Fr(f32),
-  /// A fixed length
-  #[serde(untagged)]
-  Unit(LengthUnit),
-}
-
-/// Represents a grid minmax()
-#[derive(Debug, Clone, Deserialize, Serialize, TS, PartialEq)]
-pub struct GridMinMaxSize {
-  /// The minimum size of the grid item
-  pub min: GridLengthUnit,
-  /// The maximum size of the grid item
-  pub max: GridLengthUnit,
-}
-
-/// Represents a grid track size
-#[derive(Debug, Clone, Deserialize, Serialize, TS, PartialEq)]
-#[serde(untagged)]
-pub enum GridTrackSize {
-  /// A minmax() track size
-  MinMax(GridMinMaxSize),
-  /// A fixed track size
-  Fixed(GridLengthUnit),
-}
-
-impl From<GridLengthUnit> for GridTrackSize {
-  fn from(length: GridLengthUnit) -> Self {
-    Self::Fixed(length)
-  }
-}
-
-/// Represents a grid repeat track
-#[derive(Debug, Clone, Deserialize, Serialize, TS, PartialEq)]
-pub struct GridRepeatTrack {
-  /// The size of the grid track
-  pub size: GridTrackSize,
-  /// The names of the grid lines
-  pub names: Vec<String>,
-}
-
-impl GridLengthUnit {
-  /// Converts the grid track size to a compact length representation.
-  pub fn to_compact_length(&self, context: &RenderContext) -> CompactLength {
-    match self {
-      GridLengthUnit::Fr(fr) => CompactLength::fr(*fr),
-      GridLengthUnit::Unit(unit) => unit.to_compact_length(context),
-    }
-  }
-}
-
-impl GridTrackSize {
-  /// Converts the grid track size to a non-repeated track sizing function.
-  pub fn to_min_max(&self, context: &RenderContext) -> TrackSizingFunction {
-    match self {
-      // SAFETY: The compact length is a valid track sizing function.
-      Self::Fixed(size) => unsafe {
-        TrackSizingFunction {
-          min: MinTrackSizingFunction::from_raw(size.to_compact_length(context)),
-          max: MaxTrackSizingFunction::from_raw(size.to_compact_length(context)),
-        }
-      },
-      Self::MinMax(min_max) => unsafe {
-        TrackSizingFunction {
-          min: MinTrackSizingFunction::from_raw(min_max.min.to_compact_length(context)),
-          max: MaxTrackSizingFunction::from_raw(min_max.max.to_compact_length(context)),
-        }
-      },
-    }
-  }
-}
-
-/// Represents a grid line placement with serde support
-#[derive(Debug, Clone, Deserialize, Serialize, TS, Default, PartialEq)]
-pub struct GridLine {
-  /// The start line placement
-  pub start: Option<GridPlacement>,
-  /// The end line placement
-  pub end: Option<GridPlacement>,
-}
-
-impl From<GridLine> for taffy::Line<taffy::GridPlacement> {
-  fn from(line: GridLine) -> Self {
-    Self {
-      start: line.start.unwrap_or_default().into(),
-      end: line.end.unwrap_or_default().into(),
-    }
-  }
-}
-
-/// Represents a grid placement with serde support
-#[derive(Debug, Clone, Deserialize, Serialize, TS, Default, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum GridPlacement {
-  /// Auto placement
-  #[default]
-  Auto,
-  /// Span count
-  Span(u16),
-  /// Line index (1-based)
-  #[serde(untagged)]
-  Line(i16),
-  #[serde(untagged)]
-  /// Named grid area
-  Named(String),
-}
-
-// Note: GridPlacement has a custom conversion due to its complex nature
-impl From<GridPlacement> for taffy::GridPlacement {
-  fn from(placement: GridPlacement) -> Self {
-    match placement {
-      GridPlacement::Auto => taffy::GridPlacement::Auto,
-      GridPlacement::Line(line) => taffy::GridPlacement::Line(line.into()),
-      GridPlacement::Span(span) => taffy::GridPlacement::Span(span),
-      GridPlacement::Named(_) => taffy::GridPlacement::Auto,
-    }
-  }
-}
-
-/// Represents a grid track repetition pattern
-#[derive(Debug, Clone, Deserialize, Serialize, TS, Copy, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum GridRepetitionCount {
-  /// Automatically fills the available space with as many tracks as possible
-  AutoFill,
-  /// Automatically fits as many tracks as possible while maintaining minimum size
-  AutoFit,
-  /// Specifies an exact number of track repetitions
-  #[serde(untagged)]
-  Count(u16),
-}
-
-impl From<GridRepetitionCount> for taffy::RepetitionCount {
-  fn from(repetition: GridRepetitionCount) -> Self {
-    match repetition {
-      GridRepetitionCount::AutoFill => taffy::RepetitionCount::AutoFill,
-      GridRepetitionCount::AutoFit => taffy::RepetitionCount::AutoFit,
-      GridRepetitionCount::Count(count) => taffy::RepetitionCount::Count(count),
-    }
-  }
-}
-
-/// Represents a track sizing function
-#[derive(Debug, Clone, Deserialize, Serialize, TS, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum GridTemplateComponent {
-  /// A single non-repeated track
-  Single(GridTrackSize),
-  /// Automatically generate grid tracks to fit the available space using the specified definite track lengths
-  /// Only valid if every track in template (not just the repetition) has a fixed size.
-  Repeat(GridRepetitionCount, Vec<GridRepeatTrack>),
-}
-
-impl GridTemplateComponent {
-  /// Converts this track sizing function to a Taffy-compatible format.
-  ///
-  /// # Arguments
-  ///
-  /// * `context` - The render context containing viewport information for unit resolution
-  ///
-  /// # Returns
-  ///
-  /// A `taffy::GridTemplateComponent` that can be used with the Taffy layout engine
-  pub fn to_taffy(&self, context: &RenderContext) -> taffy::GridTemplateComponent<String> {
-    match self {
-      Self::Single(track_size) => {
-        taffy::GridTemplateComponent::Single(track_size.to_min_max(context))
-      }
-      Self::Repeat(repetition, tracks) => {
-        let track_sizes = tracks
-          .iter()
-          .map(|track| track.size.to_min_max(context))
-          .collect();
-        let line_names = tracks.iter().map(|track| track.names.clone()).collect();
-
-        taffy::GridTemplateComponent::Repeat(GridTemplateRepetition {
-          count: (*repetition).into(),
-          tracks: track_sizes,
-          line_names,
-        })
-      }
     }
   }
 }
